Split out ESPRangeView to separate file
